+++ R/callDATIMapi.R
@@ -7,12 +7,16 @@
 #' default will not try again
 #' @param timeout how long should a reponse be waited for
 #' @param api_version defaults to current but can pass in version number
+#' @param verbose return raw content with data
+#' @param quiet Echo the URL which is called to the console.
 #' @return Result of DATIM API query returned as named list.
 #'
 api_get <- function(path,
                     d2_session,
                     retry = 1, timeout = 60,
-                    api_version = NULL) {
+                    api_version = NULL,
+                    verbose = FALSE,
+                    quiet = TRUE) {
 
   base_url <- d2_session$base_url
   handle <- d2_session$handle
@@ -76,7 +80,10 @@ api_get <- function(path,
 
   # removes whitespace
   url <- gsub(" ", "", url)
-  print(url)
+  if (!quiet) {
+    print(url)
+  }
+
   # retry api get block, only retries if reponse code not in 400s
   i <- 1
   response_code <- 5
@@ -129,10 +136,16 @@ api_get <- function(path,
   }
 
   # extract text response from api response
-  resp <- jsonlite::fromJSON(httr::content(resp, as = "text"),
+  content <- jsonlite::fromJSON(httr::content(resp, as = "text"),
     simplifyDataFrame = TRUE,
     flatten = TRUE
   )
 
-  return(resp)
+  if (verbose) {
+    return(list("data" = content, "api_responses" = resp))
+  } else {
+    return(content)
+  }
+
+
 }