+++ R/getMetadataEndpoint.R
@@ -49,57 +49,60 @@ duplicateResponse <- function(resp, expand, by) {
 #' @param d2_session the d2Session object, default is "d2_default_session",
 #' it will be made upon logining in to datim with loginToDATIM
 #' @param retry the number of times to try the call
+#' @param verbose return raw content with data
 #' @return the metadata response in json format and flattened
 #' @usage
 #'
-#' .getMetadataEndpoint(end_point, values, by, fields, d2_session, retry)
+#' .getMetadataEndpoint(end_point, values, by, fields, d2_session, retry, verbose)
 #'
-#' getCategories(values, by, fields, d2_session, retry)
+#' getCategories(values, by, fields, d2_session, retry, verbose)
 #'
-#' getCatCombos(values, by, fields, d2_session, retry)
+#' getCatCombos(values, by, fields, d2_session, retry, verbose)
 #'
-#' getCatOptionCombos(values, by, fields, d2_session, retry)
+#' getCatOptionCombos(values, by, fields, d2_session, retry, verbose)
 #'
-#' getCatOptionGroupSets(values, by, fields, d2_session, retry)
+#' getCatOptionGroupSets(values, by, fields, d2_session, retry, verbose)
 #'
-#' getCatOptionGroups(values, by, fields, d2_session, retry)
+#' getCatOptionGroups(values, by, fields, d2_session, retry, verbose)
 #'
-#' getCatOptions(values, by, fields, d2_session, retry)
+#' getCatOptions(values, by, fields, d2_session, retry, verbose)
 #'
-#' getDataElementGroupSets(values, by, fields, d2_session, retry)
+#' getDataElementGroupSets(values, by, fields, d2_session, retry, verbose)
 #'
-#' getDataElementGroups(values, by, fields, d2_session, retry)
+#' getDataElementGroups(values, by, fields, d2_session, retry, verbose)
 #'
-#' getDataElements(values, by, fields, d2_session, retry)
+#' getDataElements(values, by, fields, d2_session, retry, verbose)
 #'
-#' getDataSets(values, by, fields, d2_session, retry)
+#' getDataSets(values, by, fields, d2_session, retry, verbose)
 #'
-#' getIndicatorGroupSets(values, by, fields, d2_session, retry)
+#' getIndicatorGroupSets(values, by, fields, d2_session, retry, verbose)
 #'
-#' getIndicatorGroups(values, by, fields, d2_session, retry)
+#' getIndicatorGroups(values, by, fields, d2_session, retry, verbose)
 #'
-#' getIndicators(values, by, fields, d2_session, retry)
+#' getIndicators(values, by, fields, d2_session, retry, verbose)
 #'
-#' getOptionGroupSets(values, by, fields, d2_session, retry)
+#' getOptionGroupSets(values, by, fields, d2_session, retry, verbose)
 #'
-#' getOptionGroups(values, by, fields, d2_session, retry)
+#' getOptionGroups(values, by, fields, d2_session, retry, verbose)
 #'
-#' getOptionSets(values, by, fields, d2_session, retry)
+#' getOptionSets(values, by, fields, d2_session, retry, verbose)
 #'
-#' getOptions(values, by, fields, d2_session, retry)
+#' getOptions(values, by, fields, d2_session, retry, verbose)
 #'
-#' getOrgUnitGroupSets(values, by, fields, d2_session, retry)
+#' getOrgUnitGroupSets(values, by, fields, d2_session, retry, verbose)
 #'
-#' getOrgUnitGroups(values, by, fields, d2_session, retry)
+#' getOrgUnitGroups(values, by, fields, d2_session, retry, verbose)
 #'
-#' getOrgUnits(values, by, fields, d2_session, retry)
+#' getOrgUnits(values, by, fields, d2_session, retry, verbose)
 #'
-#' getDimensions(values, by, fields, d2_session, retry)
+#' getDimensions(values, by, fields, d2_session, retry, verbose)
 #'
 .getMetadataEndpoint <- function(end_point, values,
                                  by = "id",
                                  fields = NULL,
-                                 d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                                 d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1,
+                                 verbose = FALSE,
+                                 quiet = TRUE) {
   see <- try(stringi::stri_extract_all_regex(fields, "\\[[^()]+\\]")[[1]], silent = TRUE)
 
   name_reduce <- NULL
@@ -164,26 +167,42 @@ duplicateResponse <- function(resp, expand, by) {
        end_point = !!end_point,
        d2_session = d2_session,
        x,
-       fields = default_fields, retry = retry
+       fields = default_fields, retry = retry, verbose = verbose, quiet = quiet
      )})
 
-   # bind the responses
-   data <- do.call("rbind", data_list)
-  } else {
-    #normal route
+    # bind the responses
+   if (verbose) {
+     data <- do.call("rbind", lapply(data_list, `[[`, 1))
+   } else {
+     data <- do.call("rbind", data_list)
+   }
+
+    if (verbose) {
+      data <- list("data" = data, "api_responses" = data_list$api_responses)
+    }
+
+  } else { #normal route
     filters <- metadataFilter(
       values = unique_values,
       property = by,
       operator = "in"
     )
 
-    # call getMetadataa single time
+    # call getMetadata single time
     data <- getMetadata(
       end_point = !!end_point,
       d2_session = d2_session,
       filters,
-      fields = default_fields, retry = retry
+      fields = default_fields,
+      retry = retry,
+      verbose = verbose,
+      quiet = quiet
     )
+  }
+
+  if (verbose) {
+    meta_data <- data$api_responses
+    data <- data$data
   }
 
   #return NULL if there is nothing to return
@@ -207,9 +226,17 @@ duplicateResponse <- function(resp, expand, by) {
   if (length(values) == 1
       && length(data) == 1
       && is.list(data)) {
-    return(data[[1]])
+    if (verbose) {
+      return(list("data" = data[[1]], "api_responses" = meta_data))
+    } else {
+      return(data[[1]])
+    }
   } else {
-    return(data)
+    if (verbose) {
+      return(list("data" = data, "api_responses" = meta_data))
+    } else {
+      return(data)
+    }
   }
 }
 
@@ -218,13 +245,18 @@ duplicateResponse <- function(resp, expand, by) {
 getCategories <- function(values,
                           by = "id",
                           fields = NULL,
-                          d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                          d2_session = dynGet("d2_default_session", inherits = TRUE),
+                          retry = 1,
+                          verbose = FALSE,
+                          quiet = TRUE) {
   .getMetadataEndpoint("categories",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
-  )
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose,
+                       quiet = quiet)
 }
 
 #' @export
@@ -232,12 +264,18 @@ getCategories <- function(values,
 getCatCombos <- function(values,
                          by = "id",
                          fields = NULL,
-                         d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                         d2_session = dynGet("d2_default_session", inherits = TRUE),
+                         retry = 1,
+                         verbose = FALSE,
+                         quiet = TRUE) {
   .getMetadataEndpoint("categoryCombos",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose,
+                       quiet = quiet
   )
 }
 
@@ -246,12 +284,17 @@ getCatCombos <- function(values,
 getCatOptionCombos <- function(values,
                                by = "id",
                                fields = NULL,
-                               d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                               d2_session = dynGet("d2_default_session", inherits = TRUE),
+                               retry = 1,
+                               verbose = FALSE,
+                               quiet = TRUE) {
   .getMetadataEndpoint("categoryOptionCombos",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose,
   )
 }
 
@@ -260,12 +303,16 @@ getCatOptionCombos <- function(values,
 getCatOptionGroupSets <- function(values,
                                   by = "id",
                                   fields = NULL,
-                                  d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                                  d2_session = dynGet("d2_default_session", inherits = TRUE),
+                                  retry = 1,
+                                  verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("categoryOptionGroupSets",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -274,12 +321,16 @@ getCatOptionGroupSets <- function(values,
 getCatOptionGroups <- function(values,
                                by = "id",
                                fields = NULL,
-                               d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                               d2_session = dynGet("d2_default_session", inherits = TRUE),
+                               retry = 1,
+                               verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("categoryOptionGroups",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -288,12 +339,16 @@ getCatOptionGroups <- function(values,
 getCatOptions <- function(values,
                           by = "id",
                           fields = NULL,
-                          d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                          d2_session = dynGet("d2_default_session", inherits = TRUE),
+                          retry = 1,
+                          verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("categoryOptions",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -302,12 +357,16 @@ getCatOptions <- function(values,
 getDataElementGroupSets <- function(values,
                                     by = "id",
                                     fields = NULL,
-                                    d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                                    d2_session = dynGet("d2_default_session", inherits = TRUE),
+                                    retry = 1,
+                                    verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("dataElementGroupSets",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -316,12 +375,16 @@ getDataElementGroupSets <- function(values,
 getDataElementGroups <- function(values,
                                  by = "id",
                                  fields = NULL,
-                                 d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                                 d2_session = dynGet("d2_default_session", inherits = TRUE),
+                                 retry = 1,
+                                 verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("dataElementGroups",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -330,12 +393,16 @@ getDataElementGroups <- function(values,
 getDataElements <- function(values,
                             by = "id",
                             fields = NULL,
-                            d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                            d2_session = dynGet("d2_default_session", inherits = TRUE),
+                            retry = 1,
+                            verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("dataElements",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -344,12 +411,16 @@ getDataElements <- function(values,
 getDataSets <- function(values,
                         by = "id",
                         fields = NULL,
-                        d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                        d2_session = dynGet("d2_default_session", inherits = TRUE),
+                        retry = 1,
+                        verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("dataSets",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -358,12 +429,16 @@ getDataSets <- function(values,
 getUserGroups <- function(values,
                         by = "id",
                         fields = NULL,
-                        d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                        d2_session = dynGet("d2_default_session", inherits = TRUE),
+                        retry = 1,
+                        verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("userGroups",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -372,12 +447,16 @@ getUserGroups <- function(values,
 getIndicatorGroupSets <- function(values,
                                   by = "id",
                                   fields = NULL,
-                                  d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                                  d2_session = dynGet("d2_default_session", inherits = TRUE),
+                                  retry = 1,
+                                  verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("indicatorGroupSets",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -386,12 +465,16 @@ getIndicatorGroupSets <- function(values,
 getIndicatorGroups <- function(values,
                                by = "id",
                                fields = NULL,
-                               d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                               d2_session = dynGet("d2_default_session", inherits = TRUE),
+                               retry = 1,
+                               verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("indicatorGroups",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -400,12 +483,16 @@ getIndicatorGroups <- function(values,
 getIndicators <- function(values,
                           by = "id",
                           fields = NULL,
-                          d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                          d2_session = dynGet("d2_default_session", inherits = TRUE),
+                          retry = 1,
+                          verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("indicators",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -414,12 +501,16 @@ getIndicators <- function(values,
 getOptionGroupSets <- function(values,
                                by = "id",
                                fields = NULL,
-                               d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                               d2_session = dynGet("d2_default_session", inherits = TRUE),
+                               retry = 1,
+                               verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("optionGroupSets",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -428,12 +519,16 @@ getOptionGroupSets <- function(values,
 getOptionGroups <- function(values,
                             by = "id",
                             fields = NULL,
-                            d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                            d2_session = dynGet("d2_default_session", inherits = TRUE),
+                            retry = 1,
+                            verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("optionGroups",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -442,12 +537,16 @@ getOptionGroups <- function(values,
 getOptionSets <- function(values,
                           by = "id",
                           fields = NULL,
-                          d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                          d2_session = dynGet("d2_default_session", inherits = TRUE),
+                          retry = 1,
+                          verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("optionSets",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -456,12 +555,16 @@ getOptionSets <- function(values,
 getOptions <- function(values,
                        by = "id",
                        fields = NULL,
-                       d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                       d2_session = dynGet("d2_default_session", inherits = TRUE),
+                       retry = 1,
+                       verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("options",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -470,12 +573,16 @@ getOptions <- function(values,
 getOrgUnitGroupSets <- function(values,
                                 by = "id",
                                 fields = NULL,
-                                d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                                d2_session = dynGet("d2_default_session", inherits = TRUE),
+                                retry = 1,
+                                verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("organisationUnitGroupSets",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                      d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -484,12 +591,16 @@ getOrgUnitGroupSets <- function(values,
 getOrgUnitGroups <- function(values,
                              by = "id",
                              fields = NULL,
-                             d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                             d2_session = dynGet("d2_default_session", inherits = TRUE),
+                             retry = 1,
+                             verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("organisationUnitGroups",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -498,12 +609,16 @@ getOrgUnitGroups <- function(values,
 getOrgUnits <- function(values,
                         by = "id",
                         fields = NULL,
-                        d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                        d2_session = dynGet("d2_default_session", inherits = TRUE),
+                        retry = 1,
+                        verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("organisationUnits",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }
 
@@ -512,11 +627,15 @@ getOrgUnits <- function(values,
 getDimensions <- function(values,
                           by = "id",
                           fields = NULL,
-                          d2_session = dynGet("d2_default_session", inherits = TRUE), retry = 1) {
+                          d2_session = dynGet("d2_default_session", inherits = TRUE),
+                          retry = 1,
+                          verbose = FALSE, quiet = TRUE) {
   .getMetadataEndpoint("dimensions",
                        values = values,
                        by = as.character(rlang::ensym(by)),
                        fields = fields,
-                       d2_session = d2_session, retry = retry
+                       d2_session = d2_session,
+                       retry = retry,
+                       verbose = verbose, quiet = quiet
   )
 }